Remove py26 support from astara-horizon
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I97cabff07d6f8c2a24d953a129c6a5dbae92406a
Related-Bug: #1519510